refactor(agents): Split up and lint agents file; change StaticAgentRegistry to be static
