+++ src/agents/index.ts
@@ -1,0 +1,6 @@
+export { Agent } from "./agent-model";
+export { StaticAgentRegistry } from "./static-agent";
+export { AgentRecord, PropertyOperation, PropertyChange } from "./agent-record";
+export { AgentReference } from "./agent-reference";
+export { InstanceAgents } from "./instance-agents";
+export { InstanceState } from "./instance-state";
