+++ src/agents/instance-state.ts
@@ -1,0 +1,9 @@
+import GameInstance from "../game-instance";
+import { Agent } from "./agent-model";
+
+export class InstanceState extends Agent {
+    constructor(game: GameInstance) {
+        super();
+        return this.register(game, 0);
+    }
+}
